refactor(ci): split workflow registry config
