+++ tuturuuu.ci.ts
@@ -1,0 +1,75 @@
+export const ci = {
+  'actions-storage-report.yaml': true,
+  'biome-check.yaml': true,
+  'branch-name-check.yaml': true,
+  'cancel-pr-runs-on-close.yaml': true,
+  'check-and-bump-versions.yaml': true,
+  'check-docs-links.yml': true,
+  'check-migration-timestamps.yml': true,
+  'check-migrations.yml': true,
+  'codecov.yaml': true,
+  'codex-plugin.yaml': true,
+  'discord-python-ci.yml': true,
+  'docker-setup-check.yaml': true,
+  // TanStack Start and Rust migration work is paused. Keep its combined CI and
+  // Cloudflare deployment workflow disabled until active maintenance resumes.
+  'rust-backend.yml': false,
+  'i18n-check.yaml': true,
+  'mobile.yaml': true,
+  'release-ai-package.yaml': true,
+  'release-apis-package.yaml': true,
+  'release-devbox-package.yaml': true,
+  'release-google-package.yaml': true,
+  'release-hooks-package.yaml': true,
+  'release-icons-package.yaml': true,
+  'release-internal-api-package.yaml': true,
+  'release-sdk-package.yaml': true,
+  'release-supabase-package.yaml': true,
+  'release-please.yaml': true,
+  'release-types-package.yaml': true,
+  'release-typescript-config-package.yaml': true,
+  'release-ui-package.yaml': true,
+  'release-utils-package.yaml': true,
+  'sdk-version-bump.yaml': true,
+  'supabase-production.yaml': true,
+  'supabase-staging.yaml': true,
+  'turbo-unit-tests.yaml': true,
+  'type-check.yaml': true,
+  // Satellite previews are trusted workflow_dispatch-only jobs and bypass
+  // affected-path gating. Only the automatic platform preview belongs here.
+  'vercel-preview-platform.yaml': true,
+  'vercel-production-ai.yaml': true,
+  'vercel-production-apps.yaml': true,
+  'vercel-production-calendar.yaml': true,
+  'vercel-production-chat.yaml': true,
+  'vercel-production-cms.yaml': true,
+  'vercel-production-drive.yaml': true,
+  'vercel-production-finance.yaml': true,
+  'vercel-production-inventory.yaml': true,
+  'vercel-production-infrastructure.yaml': true,
+  'vercel-production-storefront.yaml': true,
+  'vercel-production-mind.yaml': true,
+  'vercel-production-nova.yaml': true,
+  'vercel-production-platform.yaml': true,
+  'vercel-production-tools.yaml': true,
+  'vercel-production-rewise.yaml': true,
+  'vercel-production-shortener.yaml': true,
+  'vercel-production-tasks.yaml': true,
+  'vercel-production-meet.yaml': true,
+  'vercel-production-learn.yaml': true,
+  'vercel-production-mail.yaml': true,
+  // Avoid automatic TanStack production builds while the migration is paused.
+  'vercel-production-tanstack-web.yaml': false,
+  'vercel-production-teach.yaml': true,
+  'vercel-production-pay.yaml': true,
+  'vercel-production-contacts.yaml': true,
+  'vercel-production-forms.yaml': true,
+  'vercel-production-git.yaml': true,
+  'vercel-production-track.yaml': true,
+  'vercel-production.yaml': true,
+  'mobile-build-ios.yaml': true,
+  'mobile-build-android.yaml': true,
+  'mobile-build-windows.yaml': true,
+  'mobile-build-macos.yaml': true,
+  'mobile-deploy-stores.yaml': true,
+};
